ai_speaking_coach/requirements.txt
streamlit==1.28.1
openai-whisper==20231117
pyaudio==0.2.13
pydub==0.25.1
numpy==1.24.3
scipy==1.11.2
librosa==0.10.0
requests==2.31.0
anthropic==0.18.1
